started off with a refactor
