+++ lib/wraith/helpers/utilities.rb
