+++ lib/wraith/save_images.rb
@@ -1,7 +1,8 @@
-require "wraith"
 require "parallel"
 require "shellwords"
-require "wraith/utilities"
+require "wraith"
+require "wraith/helpers/capture_options"
+require "wraith/helpers/save_metadata"
 
 class Wraith::SaveImages
   attr_reader :wraith, :history, :meta
@@ -61,15 +62,6 @@ class Wraith::SaveImages
     width
   end
 
-  def capture_page_image(browser, url, width, file_name, selector, global_before_capture, path_before_capture)
-
-    command = "#{browser} #{wraith.phantomjs_options} '#{wraith.snap_file}' '#{url}' \"#{width}\" '#{file_name}' '#{selector}' '#{global_before_capture}' '#{path_before_capture}'"
-
-    # @TODO - uncomment the following line when we add a verbose mode
-    # puts command
-    run_command command
-  end
-
   def run_command(command)
     output = []
     IO.popen(command).each do |line|
@@ -82,7 +74,8 @@ class Wraith::SaveImages
   def parallel_task(jobs)
     Parallel.each(jobs, :in_threads => 8) do |_label, _path, width, url, filename, selector, global_before_capture, path_before_capture|
       begin
-        attempt_image_capture(width, url, filename, selector, global_before_capture, path_before_capture, 5)
+        command = construct_command(width, url, filename, selector, global_before_capture, path_before_capture)
+        attempt_image_capture command
       rescue => e
         puts e
         create_invalid_image(filename, width)
@@ -90,9 +83,16 @@ class Wraith::SaveImages
     end
   end
 
-  def attempt_image_capture(width, url, filename, selector, global_before_capture, path_before_capture, max_attempts)
+  def construct_command(width, url, file_name, selector, global_before_capture, path_before_capture)
+    capture_page_image = "#{meta.engine} #{wraith.phantomjs_options} '#{wraith.snap_file}' '#{url}' \"#{width}\" '#{file_name}' '#{selector}' '#{global_before_capture}' '#{path_before_capture}'"
+    # @TODO - uncomment the following line when we add a verbose mode
+    # puts capture_page_image
+    return capture_page_image
+  end
+
+  def attempt_image_capture(capture_page_image, max_attempts = 5)
     max_attempts.times do |i|
-      capture_page_image meta.engine, url, width, filename, selector, global_before_capture, path_before_capture
+      run_command capture_page_image
 
       if wraith.resize
         return # @TODO - need to check if the image was generated, as per the reload example below
@@ -116,87 +116,5 @@ class Wraith::SaveImages
 
   def set_image_width(image, width)
     `convert #{image.shellescape} -background none -extent #{width}x0 #{image.shellescape}`
-  end
-end
-
-class CaptureOptions
-  attr_reader :options, :wraith
-
-  def initialize(options, wraith)
-    @options = options
-    @wraith = wraith
-  end
-
-  def path
-    has_casper(options)
-  end
-
-  def selector
-    options["selector"] || "body"
-  end
-
-  def resize
-    # path level, or YAML-file level `resize_or_reload` property value
-    if @options["resize_or_reload"]
-      (@options["resize_or_reload"] == 'resize')
-    else
-      @wraith.resize
-    end
-  end
-
-  def before_capture
-    @options["before_capture"] ? convert_to_absolute(@options["before_capture"]) : "false"
-  end
-
-  def base_url
-    base_urls(path)
-  end
-
-  def compare_url
-    compare_urls(path)
-  end
-
-  def base_urls(path)
-    wraith.base_domain + path unless wraith.base_domain.nil?
-  end
-
-  def compare_urls(path)
-    wraith.comp_domain + path unless wraith.comp_domain.nil?
-  end
-
-  def has_casper(options)
-    options["path"] ? options["path"] : options
-  end
-end
-
-class SaveMetadata
-  attr_reader :wraith, :history
-
-  def initialize(config, history)
-    @wraith = config
-    @history = history
-  end
-
-  def history_label
-    history ? "_latest" : ""
-  end
-
-  def file_names(width, label, domain_label)
-    if width.kind_of? Array
-      width = 'MULTI'
-    end
-    "#{wraith.directory}/#{label}/#{width}_#{engine}_#{domain_label}.png"
-  end
-
-  def base_label
-    "#{wraith.base_domain_label}#{history_label}"
-  end
-
-  def compare_label
-    "#{wraith.comp_domain_label}#{history_label}"
-  end
-
-  def engine
-    wraith.engine
   end
 end
